requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
duckduckgo-search
supabase
python-dotenv